+++ CHANGELOG.rst
@@ -5,6 +5,14 @@ Release Notes
 =============
 
 
+0.9.2 (07-July-2026)
+====================
+
+- Improved handling of edge cases in bounding polygon calculations for
+  catalogs in the ``WCSImage`` class. Fixed bugs in how minimal area bounding
+  rectangles are computed. [#256]
+
+
 0.9.1 (30-June-2026)
 ====================